// src/generated/client/models/DynamicPageResourceSwitch.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Muisti API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    DynamicPageResourceDataSource,
    DynamicPageResourceDataSourceFromJSON,
    DynamicPageResourceDataSourceFromJSONTyped,
    DynamicPageResourceDataSourceToJSON,
    DynamicPageResourceSwitchWhen,
    DynamicPageResourceSwitchWhenFromJSON,
    DynamicPageResourceSwitchWhenFromJSONTyped,
    DynamicPageResourceSwitchWhenToJSON,
} from './';

/**
 * 
 * @export
 * @interface DynamicPageResourceSwitch
 */
export interface DynamicPageResourceSwitch {
    /**
     * 
     * @type {DynamicPageResourceDataSource}
     * @memberof DynamicPageResourceSwitch
     */
    dataSource: DynamicPageResourceDataSource;
    /**
     * 
     * @type {string}
     * @memberof DynamicPageResourceSwitch
     */
    key?: string;
    /**
     * 
     * @type {Array<DynamicPageResourceSwitchWhen>}
     * @memberof DynamicPageResourceSwitch
     */
    when?: Array<DynamicPageResourceSwitchWhen>;
}

export function DynamicPageResourceSwitchFromJSON(json: any): DynamicPageResourceSwitch {
    return DynamicPageResourceSwitchFromJSONTyped(json, false);
}

export function DynamicPageResourceSwitchFromJSONTyped(json: any, ignoreDiscriminator: boolean): DynamicPageResourceSwitch {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'dataSource': DynamicPageResourceDataSourceFromJSON(json['dataSource']),
        'key': !exists(json, 'key') ? undefined : json['key'],
        'when': !exists(json, 'when') ? undefined : ((json['when'] as Array<any>).map(DynamicPageResourceSwitchWhenFromJSON)),
    };
}

export function DynamicPageResourceSwitchToJSON(value?: DynamicPageResourceSwitch | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'dataSource': DynamicPageResourceDataSourceToJSON(value.dataSource),
        'key': value.key,
        'when': value.when === undefined ? undefined : ((value.when as Array<any>).map(DynamicPageResourceSwitchWhenToJSON)),
    };
}